#344 fixed sonarcloud bug and code smells

+++ backend/src/main/java/com/capstone/moneytree/service/impl/DefaultUserService.java
@@ -488,7 +488,7 @@ public class DefaultUserService implements UserService {
         float index = 0;
         for (int i = 0; i < usernames.size(); i++) {
             if (usernames.get(i).equals(username)) {
-                index = i+1;
+                index = (float) i+1;
                 break;
             }
         }
